+++ lib/material/list.rb
@@ -1,19 +1,16 @@
 # frozen_string_literal: true
 
-require_relative "concerns/components"
-require_relative "concerns/icon"
-require_relative "concerns/text"
-
 module Material
   class List < Collectible::CollectionBase
     include Material::Components
-    include Material::Icon
+    include Material::Display
     include Material::Text
     include Material::Site
     include Material::For
     include Material::Pagination
     include Material::Mount
     include Material::Collection
+    include Material::Format
 
     class << self
       def for(object)