+++ spec/material/base_spec.rb
@@ -7,10 +7,13 @@ RSpec.describe Material::Base, type: :material do
 
   it { is_expected.to include_module Material::Components }
   it { is_expected.to include_module Material::Core }
-  it { is_expected.to include_module Material::Icon }
+  it { is_expected.to include_module Material::Display }
   it { is_expected.to include_module Material::Text }
   it { is_expected.to include_module Material::Site }
   it { is_expected.to include_module Material::For }
+  it { is_expected.to include_module Material::Format }
+
+  it { is_expected.to have_material_component :list_item_style }
 
   describe ".for" do
     include_context "with an example material"