+++ spec/material/list_spec.rb
@@ -6,13 +6,14 @@ RSpec.describe Material::List, type: :material do
   it { is_expected.to inherit_from Collectible::CollectionBase }
 
   it { is_expected.to include_module Material::Components }
-  it { is_expected.to include_module Material::Icon }
+  it { is_expected.to include_module Material::Display }
   it { is_expected.to include_module Material::Text }
   it { is_expected.to include_module Material::Site }
   it { is_expected.to include_module Material::For }
   it { is_expected.to include_module Material::Pagination }
   it { is_expected.to include_module Material::Mount }
   it { is_expected.to include_module Material::Collection }
+  it { is_expected.to include_module Material::Format }
 
   describe ".for" do
     include_context "with an example list"